+++ tests/test_requests.py
@@ -93,3 +93,16 @@ def test_requests_log_insert_to_es(
     doc_id = log(requests_request, requests_response)
     actual_full_request_log = get_from_es(doc_id)
     assert actual_full_request_log == serialized_full_request_log
+
+
+@pytest.mark.skipif(
+    not config.ELASTIC_SEARCH_RUN_TESTS,
+    reason="Set ELASTIC_SEARCH_RUN_TESTS env to enable Elastic Search tests",
+)
+def test_requests_logging_with_duration(
+        freezer, test_es_index, requests_request, requests_response, serialized_full_request_log,
+):
+    duration = 1.2
+    doc_id = log(requests_request, requests_response, duration=duration)
+    actual_full_request_log = get_from_es(doc_id)
+    assert actual_full_request_log["stats"]["duration"] == duration